app-python/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.25.2
vectorbt==0.25.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
# supabase==2.0.2  # Removido - usando SQLite local
python-multipart==0.0.6
